refactor(onboarding): Create TransferCreditGallery component
Moved implementation of TransferCredit from Onboarding_Pages/pg_3.tsx to
components/onboarding/TransferCreditGallery.tsx to reduce complexity of
Onboarding_Pages/pg_3.tsx
